+++ src/Support/helpers.php
@@ -9,15 +9,35 @@ if (!function_exists('readtime')) {
      * @param string|null       $locale
      * @param array             $options
      *
-     * @return string
+     * @return string|null
      */
     function readtime(
         $content,
         string $locale = null,
         array $options = []
-    ): string
+    ): ?string
     {
         return ReadTime::parse($content, $options)
             ->get($locale);
+    }
+}
+
+if (!function_exists('pull_first_substr')) {
+
+    /**
+     * @param string $string
+     * @param string $separator
+     *
+     * @return string
+     */
+    function pull_first_substr(string &$string, string $separator = '.'): string
+    {
+        $arr = explode($separator, $string);
+
+        $value = array_shift($arr);
+
+        $string = implode('.', $arr);
+
+        return $value;
     }
 }