+++ tests/ExampleTest.php
@@ -1,20 +1,31 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Leshkens\LaravelReadtime\Tests;
 
 use Orchestra\Testbench\TestCase;
-use Leshkens\LaravelReadtime\LaravelReadtimeServiceProvider;
+use Leshkens\LaravelReadtime\Providers\ReadtimeServiceProvider;
 
+/**
+ * Class ExampleTest
+ *
+ * @package Leshkens\LaravelReadtime\Tests
+ */
 class ExampleTest extends TestCase
 {
-
-    protected function getPackageProviders($app)
+    /**
+     * @param $app
+     *
+     * @return string[]
+     */
+    protected function getPackageProviders($app): array
     {
-        return [LaravelReadtimeServiceProvider::class];
+        return [ReadtimeServiceProvider::class];
     }
-    
+
     /** @test */
-    public function true_is_true()
+    public function true_is_true(): void
     {
         $this->assertTrue(true);
     }