+++ CHANGELOG.md
@@ -1,7 +1,17 @@
 # Changelog
 
-All notable changes to `laravel-readtime` will be documented in this file
+## 1.1.0 - 2021-04-01
 
-## 1.0.0 - 201X-XX-XX
+- Support for "dot" notation in the source, in the model
+- Refactoring `getReadTimeAttribute()` in `HasReadTime` trait
+- Small fixes
+
+
+## 1.0.1 - 2021-03-31
+
+- Small fixes
+
+
+## 1.0.0 - 2021-03-31
 
 - initial release